[FMHA] rename dualwave_common.py to flash_attn_utils.py
- Update imports in gfx950, fp8, generic, and interface modules
- Import scf_if_dispatch from flash_attn_utils in generic
